Fixed maim not detecting a format input, fixed old glm not wanting to cast from glm::vec4 to glm::ivec4

+++ src/x.cpp
@@ -7,7 +7,7 @@ TmpXError(Display * d, XErrorEvent * ev) {
     return 0;
 }
 /*
-glm::vec4 getWindowGeometry( X11* x11, Window win ) {
+glm::ivec4 getWindowGeometry( X11* x11, Window win ) {
     XWindowAttributes attr;
     XGetWindowAttributes( x11->display, win, &attr );
     unsigned int width = attr.width;           
@@ -19,7 +19,7 @@ glm::vec4 getWindowGeometry( X11* x11, Window win ) {
     return glm::vec4( x, y, width, height );
 } 
 */
-glm::vec4 getWindowGeometry( X11* x11, Window win ) {
+glm::ivec4 getWindowGeometry( X11* x11, Window win ) {
     XWindowAttributes attr;         
     XGetWindowAttributes( x11->display, win, &attr );
     unsigned int width = attr.width;
