+++ src/x.hpp
@@ -51,7 +51,7 @@ public:
     void freeCRTCS( std::vector<XRRCrtcInfo*> monitors );
 };
 
-glm::vec4 getWindowGeometry( X11* x11, Window win );
-//glm::vec4 getWindowGeometryWithoutBorder( X11* x11, Window win );
+glm::ivec4 getWindowGeometry( X11* x11, Window win );
+//glm::ivec4 getWindowGeometryWithoutBorder( X11* x11, Window win );
 
 #endif
